[fix] Upgrade internals of Produce, fixing inconsistent distribution
Produce now has predictable consumption probabilities, for functions of any
arity. For any "leaf" of a partially consumed input, there is a ~1/2^N chance
that it will be expanded by N recursive consumptions. The actual math for finite
data structures with non-regular branching patterns is slightly more
complicated, left as an exercise to the reader. :)