[fix] Resolve bug in derived Observe for 1-constructor 1-field types
Observe was incorrectly derived for a type if it:

1. derived Observe via -XDeriveAnyClass or an empty instance declaration, and
2. had only 1 constructor, and
3. had only 1 field in that constructor.

Steps to reproduce:

```
data D = C ()
  deriving stock (GHC.Generic)
  deriving anyclass (Generic, HasDatatypeInfo, Consume, Observe)
```

Evaluating the following *should* yield the demand `C _`:

```
fst $ observe1 (`seq` ()) id (C ())
```

However, it yielded the demand `_`, which is obviously incorrect, since the true
demand we placed was a demand to whnf.

This issue was due to insufficient strictness in the generic implementations of
`projectD` and `embedD`. In particular, they only forced their value or demand
argument when that argument had more than one constructor or field. This
happened because Generics.SOP is maximally lazy in deriving its own `to`/`from`
functions through the equally lazy functions in GHC.Generics. Thus, values with
only one constructor and only one field were forced too late.

Because we rely on the invariant that evaluating a constructor of an entangled
demand evaluates precisely the corresponding constructor of the data it is
associated with, this out-of-sync laziness caused us to report smaller demands
than were actually exerted.