[refactor] Extract recursion-schemes-ish things to Shaped class
We have realized that the previously-named Observe class did not contain
anything particularly specific to observation -- it just allowed us to manipulate
interleaved functors inside of datatypes, ala recursion-schemes but at a higher
kind. As such, this commit refactors the previously-named Observe class into a
Shaped class, and updates all modules to reflect this change.